refactor: change folder location and delete unnessary types
