+++ src/components/Form/Form.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 import { FieldValues, FormProvider, SubmitHandler, useForm } from 'react-hook-form';
 import { TFormList } from 'data/types';
 
-import Firstname from './Elements/Firstname';
-import Secondname from './Elements/Secondname';
-import Email from './Elements/Email';
-import Gender from './Elements/Gender';
-import Country from './Elements/Country';
-import Birthday from './Elements/Birthday';
-import UploadPhoto from './Elements/Photo';
-import Agree from './Elements/Agree';
+import Firstname from '../FormItems/Firstname';
+import Secondname from '../FormItems/Secondname';
+import Email from '../FormItems/Email';
+import Gender from '../FormItems/Gender';
+import Country from '../FormItems/Country';
+import Birthday from '../FormItems/Birthday';
+import UploadPhoto from '../FormItems/Photo';
+import Agree from '../FormItems/Agree';
 
 export type createdCardT = {
   createdCard: (data: TFormList) => void;
