+++ src/components/Modal/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ModalWindow() {
+  return <div>Тю-тю-модальное окно</div>;
+}
+
+export default ModalWindow;
