+++ src/data/types.ts
@@ -1,11 +1,3 @@
-export type Unique = {
-  [index: string]: string | number;
-};
-
-export type State = {
-  [index: string]: string | number;
-};
-
 export type TFormList = {
   firstname: string;
   secondname: string;
@@ -14,20 +6,6 @@ export type TFormList = {
   gender: string;
   birthday: string;
   photo: string;
-};
-
-export type TCardList = {
-  key: number;
-  name: string;
-  calories: string;
-  protein: string;
-  fiber: string;
-  portion: string;
-  photo: string;
-};
-
-export type TFormListProps = {
-  props?: string;
 };
 
 export type List = {
